adding more to subcategory, validation for admin

+++ src/main/java/com/example/technomarket/services/CategoryService.java
@@ -3,8 +3,12 @@ package com.example.technomarket.services;
 import com.example.technomarket.model.dto.categoryDTOs.CategoryWithNameOnlyDTO;
 import com.example.technomarket.model.dto.categoryDTOs.CategoryWithNewNameDTO;
 import com.example.technomarket.model.exceptions.BadRequestException;
+import com.example.technomarket.model.exceptions.UnauthorizedException;
 import com.example.technomarket.model.pojo.Category;
+import com.example.technomarket.model.pojo.SubCategory;
 import com.example.technomarket.model.repository.CategoryRepository;
+import com.example.technomarket.model.repository.SubcategoryRepository;
+import com.example.technomarket.util.CurrentUser;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,7 +22,12 @@ public class CategoryService {
     @Autowired
     private CategoryRepository categoryRepository;
     @Autowired
-    ModelMapper modelMapper;
+    private SubcategoryRepository subcategoryRepository;
+    @Autowired
+    private ModelMapper modelMapper;
+    @Autowired
+    private CurrentUser currentUser;
+
 
     public List<CategoryWithNameOnlyDTO> showAllCategories(){
         List<Category> category = categoryRepository.findAll();
@@ -26,6 +35,9 @@ public class CategoryService {
     }
 
     public void addCategoryToList(CategoryWithNameOnlyDTO category) {
+        if(!currentUser.isAdmin()){
+            throw new UnauthorizedException("Method not allowed!");
+        }
         Optional<Category> optionalCategory = categoryRepository.findCategoryByName(category.getCategoryName());
         if(category.getCategoryName() != null && optionalCategory.isEmpty()){
             Category c = modelMapper.map(category, Category.class);
@@ -37,8 +49,13 @@ public class CategoryService {
     }
 
     public void deleteCategory(long id) {
+        if(!currentUser.isAdmin()){
+            throw new UnauthorizedException("Method not allowed!");
+        }
         Optional<Category> categoryOptional = categoryRepository.findCategoryByCategoryId(id);
         if(categoryOptional.isPresent()){
+            List<SubCategory> subCategories = subcategoryRepository.findAllSubCategoryByCategory(categoryOptional.get());
+            subcategoryRepository.deleteAll(subCategories);
             Category category = modelMapper.map(categoryOptional, Category.class);
             categoryRepository.delete(category);
         }
@@ -48,6 +65,9 @@ public class CategoryService {
     }
 
     public void updateCategory(long cid, CategoryWithNewNameDTO category) {
+        if(!currentUser.isAdmin()){
+            throw new UnauthorizedException("Method not allowed!");
+        }
         Optional<Category> categoryOptional= categoryRepository.findCategoryByCategoryId(cid);
         if(categoryOptional.isPresent()){
             Category category1 = categoryOptional.get();
